chore(node): refactor application generator for more clarity
