+++ packages/node/src/generators/application/lib/normalized-schema.ts
@@ -1,0 +1,9 @@
+import { Schema } from '../schema';
+
+export interface NormalizedSchema extends Omit<Schema, 'useTsSolution'> {
+  appProjectRoot: string;
+  parsedTags: string[];
+  outputPath: string;
+  importPath: string;
+  isUsingTsSolutionConfig: boolean;
+}
